1.5.2) Przeorganizowano aplikację tak, aby usunąć plik klasy publicznej Node
